+++ src/modules/status/status.service.ts
@@ -32,11 +32,10 @@ import { ConfigService } from '../../core/config/config.service.js'
 import { HomebridgeIpcService } from '../../core/homebridge-ipc/homebridge-ipc.service.js'
 import { Logger } from '../../core/logger/logger.service.js'
 import { isNodeV24SupportedArchitecture } from '../../core/node-version.constants.js'
+import { DockerRelease, DockerReleaseInfo } from '../platform-tools/docker/docker.interfaces.js'
 import { PluginsService } from '../plugins/plugins.service.js'
 import { ServerService } from '../server/server.service.js'
 import {
-  DockerRelease,
-  DockerReleaseInfo,
   HomebridgeStatsResponse,
   HomebridgeStatus,
   HomebridgeStatusUpdate,